fixed some mistakes

+++ contracts/interfaces/ICurveMinter.sol
@@ -36,7 +36,7 @@ interface ICurveMinter {
 
     function remove_liquidity(uint256 _amount, uint256[3] calldata amounts, bool use_underlying) external;
 
-    function exchange(int128 from, int128 to, uint256 _from_amount, uint256 _min_to_amount) external returns (uint256);
+    function exchange(uint256 from, uint256 to, uint256 _from_amount, uint256 _min_to_amount) external returns (uint256);
 
     function exchange_underlying(int128 i, int128 j, uint256 dx, uint256 min_dy) external;
